fix(db): sync_run.id を BIGSERIAL に修正 + CI で integrationTest 実行
CI の Testcontainers IT 失敗の根本原因:
  Schema validation: wrong column type encountered in column [id] in table
  [sync_run]; found [serial (Types#INTEGER)], but expecting [bigint (Types#BIGINT)]

- V1__init.sql: sync_run.id を SERIAL -> BIGSERIAL (Domain SyncRun.id: Long? と
  整合)。log テーブルなので 64-bit が妥当
- docs/01-design.md §4.2: sync_run の id を BIGSERIAL PK に修正
- .github/workflows/ci.yml: integrationTest ステップを追加し Testcontainers IT
  も CI で常時実行する

検証 (local postgres:16 で V1 を実行):
- sync_run.id: bigint, default nextval('sync_run_id_seq')
- 他カラムは従来通り

+++ backend/src/main/resources/db/migration/V1__init.sql
@@ -28,7 +28,8 @@ COMMENT ON COLUMN sync_state.tickets_total         IS 'Qdrant 上のユニーク
 
 
 CREATE TABLE sync_run (
-    id                SERIAL       PRIMARY KEY,
+    -- BIGSERIAL: log テーブルなので 32-bit 上限を回避し SyncRunEntity (Long) と合わせる
+    id                BIGSERIAL    PRIMARY KEY,
     kind              VARCHAR(32)  NOT NULL,
     started_at        TIMESTAMPTZ  NOT NULL,
     finished_at       TIMESTAMPTZ,
